+++ CHANGELOG.md
@@ -53,6 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix Data Browser regression where data rows become invalid after switching between relations
   ([#993](https://github.com/feldera/feldera/issues/993))
   ([#999](https://github.com/feldera/feldera/issues/999))
+- Refactor list of tables and views in breadcrumbs to an autocomplete combobox
+  ([997](https://github.com/feldera/feldera/issues/997))
 
 ## [0.3.0] - 2023-11-07